somes changes

+++ LinkedinClone/Repositories/JobPostRepository.cs
@@ -1,4 +1,5 @@
 ﻿using LinkedinClone.Data;
+using LinkedinClone.Dtos;
 using LinkedinClone.Models;
 using LinkedinClone.Repositories.Interfaces;
 using Microsoft.EntityFrameworkCore;
@@ -22,14 +23,28 @@ namespace LinkedinClone.Repositories
             return create.Entity;
         }
 
-        public async Task DeleteJobPost(int id)
+        public async Task<ResponseDto> DeleteJobPost(int id, string UserId)
         {
             var jobPostid = await _context.JobPosts.FirstOrDefaultAsync(x=>x.Id == id);
+            if(jobPostid.User.Id != UserId)
+            {
+                return new ResponseDto()
+                {
+                    Message = "You are not the creator of this job-post",
+                    StatusCode = StatusCodes.Status401Unauthorized
+                };
+            }
             if (jobPostid != null)
             {
                 _context.JobPosts.Remove(jobPostid);
                 await _context.SaveChangesAsync();
             }
+
+            return new ResponseDto()
+            {
+                Message = "Job-Post Deleted",
+                StatusCode = StatusCodes.Status200OK
+            };
         }
         
         public async Task<List<JobPost>> Get()
@@ -44,9 +59,20 @@ namespace LinkedinClone.Repositories
             return jobId;
         }
 
-        public async Task<JobPost> UpdateJobPost(JobPost jobPost, int id)
+        public async Task<ResponseDto> UpdateJobPost(JobPost jobPost, int id, string UserId)
         {
             var jobId = await _context.JobPosts.FirstOrDefaultAsync(x => x.Id == id);
+
+            if(jobId.UserId != UserId)
+            {
+                return new ResponseDto()
+                {
+                    Message = "Yo are not the creator of this job-post",
+                    StatusCode = StatusCodes.Status401Unauthorized
+                };
+            }
+
+
             if(jobId != null)
             {
                 jobId.Title = jobPost.Title;
@@ -54,7 +80,12 @@ namespace LinkedinClone.Repositories
 
                 await _context.SaveChangesAsync();
             }
-            return jobId;
+            return new ResponseDto()
+            {
+                Message = "Job-Post Updated",
+                StatusCode = StatusCodes.Status200OK,
+
+            };
         }
     }
 }
